wip packaging test
Creates a role user for elasticsearch on windows and chowns the archive
installation to that user. This works on 2016 but not on 2012r2 because
it has an older version of powershell and WMF that doesn't support those
user operations.

It can probably be achieved by doing net user /create and net username
/delete, whatever that command is called. Not sure if that creates a
local or domain user.

It can also be achieved by updating WMF and powershell to 5.1 on server
2012r2, which can be done by downloading and running an msupdate package
instructions here
(https://docs.microsoft.com/en-us/powershell/wmf/5.1/install-configure#install-wmf-51-for-windows-server-2012-r2-windows-server-2012-and-windows-81)
 but it seems like that requires a restart, not sure how that works
 during provisioning